requirements.txt
numpy==1.23.0
xgboost==1.7.1
scikit-learn==0.24.2
pandas==1.5.3
pyyaml==6.0
gensim==4.3.0
scipy==1.10.1
sympy==1.13.3
tqdm==4.67.0
filelock==3.16.1
transformers==4.31.0
protobuf==3.20.3
spacy==3.4.2
attrs==24.2.0
nltk==3.9.1